Add sosreport log collecting support
[01].Add support for opensuse tumbleweed/leap,oracle linux,fedora and rhel.
[02].Use "nmap -T4" instead of "-T5" to avoid missing ip address by using
insane host discovery.
[03].Add support for sosreport log collecting on host/guest and via ssh or
serial console.

+++ data/virt_autotest/virt_logs_collector.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 set -o pipefail
+shopt -s nocasematch
 
 #Obtain SLES release version and service pack level
 function get_sles_release() {
@@ -53,8 +54,8 @@ function collect_logs_via_guest_console() {
         touch ${expfile}
         chmod 777 ${expfile}
         local hypervisor=`get_sles_hypervisor`
-
 cat <<EOF > ${expfile}
+#!/usr/bin/expect
 set hypervisor [lindex \$argv 0]
 set guest_domain [lindex \$argv 1]
 set guest_transformed [lindex \$argv 2]
@@ -76,7 +77,7 @@ while { \${retry_times} > 0 } {
       expect  {
          -nocase "escape character" {send "\r\r"; exp_continue -continue_timer}
          -nocase "password:" {send "\r"}
-         " #" {send "exit\r"}
+         -re "( |\\\])#" {send "exit\r"}
          -nocase "mistake|wrong|fault|error|fail|exception|not*found|timed*out" {puts "Can not establish virsh console to \${guest_domain}\r"; set ret_result 1}
       }
       expect {
@@ -87,22 +88,26 @@ while { \${retry_times} > 0 } {
 
       if { \${ret_result} == 0 } {
          set timeout 1200
-         expect "~ #"
+         expect -re "~( |\\\])#"
          send "mkdir -p \${logs_folder};export time_stamp=\`date \'+%Y%m%d%H%M%S\'\`\r"
-         expect "~ #"
+         expect -re "~( |\\\])#"
          if { \${extra_logs} == {support_config} } {
             send "rm -f -r \${logs_folder}/*supportconfig*\r"
             send "supportconfig -y -A -t \${logs_folder} -B guest_\${guest_transformed}_supportconfig_\\\${time_stamp}\r"
          }
-         if { \${extra_logs} != {support_config} && \${extra_logs} != "" } {
+         if { \${extra_logs} == {sos_report} } {
+            send "rm -f -r \${logs_folder}/*sosreport*\r"
+            send "sosreport --batch --debug -v --alloptions --all-logs -z xz --tmp-dir \${logs_folder}\r"
+         }
+         if { \${extra_logs} != {support_config} && \${extra_logs} != {sos_report} && \${extra_logs} != "" } {
             send "rm -f -r \${logs_folder}/*extra_logs*\r"
             send "mkdir -p \${logs_folder}/guest_\${guest_transformed}_extra_logs;cp --parent -r -f \${extra_logs} \${logs_folder}/guest_\${guest_transformed}_extra_logs\r"
          }
          expect {
-            "~ #" {send "export ret=\\\$?;cd /tmp;(if \[\[ \\\$ret -ne 0 ]];then echo \${fail_string};else echo \${pass_string};fi)\r"; exp_continue -continue_timer}
+            -re "~( |\\\])#" {send "export ret=\\\$?;cd /tmp;(if \[\[ \\\$ret -ne 0 ]];then echo \${fail_string};else echo \${pass_string};fi)\r"; exp_continue -continue_timer}
             "sad_to_fail" {puts "Can not collect \${extra_logs} from \${guest_domain}\r"; set ret_result 1;  exp_continue -continue_timer}
             "glad_go_pass" {puts "Finished collecting \${extra_logs} from \${guest_domain}\r"; set ret_result 0;  exp_continue -continue_timer}
-            "tmp #" {send "cd ~\r"; send "exit\r"}
+            -re "tmp( |\\\])#" {send "cd ~\r"; send "exit\r"}
          }
       }
 
@@ -144,10 +149,10 @@ EOF
 	fi
 }
 
-#This function supports collecting supportconfig from both host and guest. The argument target_type will be given 'host' or 'guest'
+#This function supports collecting supportconfig or sosreport from both host and guest. The argument target_type will be given 'host' or 'guest'
 #Will resort to guest virsh console if collecting from guest ssh failed. Collecting logs from host only supports local host
-#Typical usage: collect_supportconfig logs_folder host or collect_supportconfig logs_folder guest guest_ip guest_domain_name
-function collect_supportconfig() {
+#Typical usage: collect_system_log_and_diagnosis logs_folder host or collect_system_log_and_diagnosis logs_folder guest guest_ip guest_domain_name
+function collect_system_log_and_diagnosis() {
 	local logs_folder=$1
 	local target_type=$2
 	local target_ipaddr=$3
@@ -172,23 +177,36 @@ function collect_supportconfig() {
 	local ret_result=128
 	local retry_times=0
 	while [[ ${retry_times} -lt 3 ]] && [[ ${ret_result} -ne 0 ]];
-	do	
-    	   local time_stamp=`date '+%Y%m%d%H%M%S'`
-	   ${sshpass_ssh_cmd} rm -f -r ${logs_folder}/*supportconfig*
-	   echo -e "${sshpass_ssh_cmd} supportconfig -y -A -t ${logs_folder} -B ${target_type}_${target_transformed}_supportconfig_${time_stamp}"
-	   ${sshpass_ssh_cmd} supportconfig -y -A -t ${logs_folder} -B ${target_type}_${target_transformed}_supportconfig_${time_stamp}
+	do
+	   if [[ ${target_type} == "host" && `cat /etc/issue` =~ oracle|rhel|red.*hat|fedora ]] || [[ ${target_type} == "guest" && ${target_transformed} =~ oracle|rhel|fedora ]];then
+	      ${sshpass_ssh_cmd} rm -f -r ${logs_folder}/*sosreport*
+	      echo -e "${sshpass_ssh_cmd} mkdir -p ${logs_folder}"
+	      echo -e "${sshpass_ssh_cmd} sosreport --batch --debug -v --alloptions --all-logs -z xz --tmp-dir ${logs_folder}"
+	      ${sshpass_ssh_cmd} mkdir -p ${logs_folder}
+	      ${sshpass_ssh_cmd} sosreport --batch --debug -v --alloptions --all-logs -z xz --tmp-dir ${logs_folder}
+	   else	   
+    	      local time_stamp=`date '+%Y%m%d%H%M%S'`
+	      ${sshpass_ssh_cmd} rm -f -r ${logs_folder}/*supportconfig*
+	      echo -e "${sshpass_ssh_cmd} supportconfig -y -A -t ${logs_folder} -B ${target_type}_${target_transformed}_supportconfig_${time_stamp}"
+	      ${sshpass_ssh_cmd} supportconfig -y -A -t ${logs_folder} -B ${target_type}_${target_transformed}_supportconfig_${time_stamp}
+	   fi
 	   ret_result=$?
 	   if [[ ${ret_result} -eq 0 ]];then
-	      echo -e "Successfully collected supportconfig from ${target_type} ${target_domain} via ssh."
+	      echo -e "Successfully collected supportconfig or sosreport from ${target_type} ${target_domain} via ssh."
 	      break
 	   fi
 	   retry_times=$((${retry_times}+1))
 	done
 
 	if [[ ${target_type} == "guest" ]] && [[ ${ret_result} -ne 0 ]];then
-	   echo -e "Can not collect supportconfig from ${target_type} ${target_domain} via ssh. Try to use guest virsh console."
-	   echo -e "collect_supportconfig_via_guest_console ${target_domain} ${logs_folder}"
-	   collect_supportconfig_via_guest_console ${target_domain} ${logs_folder}
+	   echo -e "Can not collect supportconfig or sosreport from ${target_type} ${target_domain} via ssh. Try to use guest virsh console."
+	   if [[ ${target_domain} =~ oracle|rhel|fedora ]];then
+	      echo -e "collect_sosreport_via_guest_console ${target_domain} ${logs_folder}"
+	      collect_sosreport_via_guest_console ${target_domain} ${logs_folder}
+	   else 
+	      echo -e "collect_supportconfig_via_guest_console ${target_domain} ${logs_folder}"
+	      collect_supportconfig_via_guest_console ${target_domain} ${logs_folder}
+	   fi
 	   if [[ $? -eq 0 ]];then
 	      return 0
 	   else
@@ -206,6 +224,17 @@ function collect_supportconfig_via_guest_console() {
         local guest_domain=$1
         local logs_folder=$2
         collect_logs_via_guest_console ${guest_domain} ${logs_folder} support_config
+        if [[ $? -eq 0 ]];then
+           return 0
+        else
+           return 1
+        fi
+}
+
+function collect_sosreport_via_guest_console() {
+        local guest_domain=$1
+        local logs_folder=$2
+        collect_logs_via_guest_console ${guest_domain} ${logs_folder} sos_report
         if [[ $? -eq 0 ]];then
            return 0
         else
@@ -405,7 +434,7 @@ fi
 
 unset guest_hash_ipaddr
 declare -a guest_hash_ipaddr=""
-guest_domain_types="sles"
+guest_domain_types="sles|opensuse|tumbleweed|leap|oracle"
 guests_inactive_array=`virsh list --inactive | grep -E "${guest_domain_types}" | awk '{print $2}'`
 guest_domains_array=`virsh list  --all | grep -E "${guest_domain_types}" | awk '{print $2}'`
 guest_macaddresses_array=""
@@ -432,7 +461,7 @@ for single_subnet in ${subnets_in_route[@]};do
     single_subnet_scan_results=${virt_logs_folder}'/nmap_subnets_scan_results/nmap_scan_'${single_subnet_transformed}'_'${scan_timestamp}
     subnets_scan_results[${subnets_scan_index}]=${single_subnet_scan_results}
     echo -e "nmap -sn $single_subnet -oX $single_subnet_scan_results" | tee -a ${virt_logs_collecor_log}
-    nmap -T5 -sn $single_subnet -oX $single_subnet_scan_results  | tee -a ${virt_logs_collecor_log}
+    nmap -T4 -sn $single_subnet -oX $single_subnet_scan_results  | tee -a ${virt_logs_collecor_log}
     subnets_scan_index=$(( ${subnets_scan_index} + 1 ))
 done
 
@@ -458,8 +487,8 @@ done
 
 #Start collecing logs from host and virtual machine
 setup_common_logs_folder ${virt_logs_folder}	
-echo -e "collect_supportconfig ${virt_logs_folder} host" | tee -a ${virt_logs_collecor_log}
-collect_supportconfig ${virt_logs_folder} host | tee -a ${virt_logs_collecor_log}
+echo -e "collect_system_log_and_diagnosis ${virt_logs_folder} host" | tee -a ${virt_logs_collecor_log}
+collect_system_log_and_diagnosis ${virt_logs_folder} host | tee -a ${virt_logs_collecor_log}
 virt_logs_collector_result=$(( ${virt_logs_collector_result} | $? ))
 echo -e "collect_extra_logs_from_host ${virt_logs_folder} ${virt_extra_logs_host}" | tee -a ${virt_logs_collecor_log}
 collect_extra_logs_from_host ${virt_logs_folder} "" ${virt_extra_logs_host} | tee -a ${virt_logs_collecor_log}
@@ -473,8 +502,8 @@ else
           if [[ ${guests_inactive_array[@]} =~ .*${guest_current}.* ]];then 
              echo -e "Virtual machine ${guest_current} in shutdown state. Skip collecting logs from it." | tee -a ${virt_logs_collecor_log}
           else
-             echo -e "collect_supportconfig ${virt_logs_folder} guest ${guest_hash_ipaddr[${guest_hash_index}]} ${guest_current}" | tee -a ${virt_logs_collecor_log}
-             collect_supportconfig ${virt_logs_folder} guest ${guest_hash_ipaddr[${guest_hash_index}]} ${guest_current} | tee -a ${virt_logs_collecor_log}
+             echo -e "collect_system_log_and_diagnosis ${virt_logs_folder} guest ${guest_hash_ipaddr[${guest_hash_index}]} ${guest_current}" | tee -a ${virt_logs_collecor_log}
+             collect_system_log_and_diagnosis ${virt_logs_folder} guest ${guest_hash_ipaddr[${guest_hash_index}]} ${guest_current} | tee -a ${virt_logs_collecor_log}
              virt_logs_collector_result=$(( ${virt_logs_collector_result} | $? ))
              echo -e "collect_extra_logs_from_guest ${virt_logs_folder} ${guest_hash_ipaddr[${guest_hash_index}]} ${guest_current}  ${virt_extra_logs_guest}" | tee -a ${virt_logs_collecor_log}
              collect_extra_logs_from_guest ${virt_logs_folder} ${guest_hash_ipaddr[${guest_hash_index}]} ${guest_current} ${virt_extra_logs_guest} | tee -a ${virt_logs_collecor_log}
